WIP on SequenceEquals
The main challenge is the actual relational implementation.
SequenceEquals only makes sense over primitive collections; these can
be compared, whereas other relational ShapedQueryExpressions cannot
(maybe possible to project the results of each side with a row number,
and then somehow compare...).

Then, ideally we want to relational logic to be able to identify
primitive collections, but there's currently no way to do that which
isn't provider-specific (we get a translated OpenJson expression).
Even when we get pre-translated, we can identify ParameterQueryRootExpression,
but what can we do for columns?