WIP
ColumnExpression isn't the only thing referencing a table - StructuralTypeProjectionExpression also does (in its TableMap).
That means TableReferenceExpression must indeed be an expression, and the cloning visitor (or the thing it uses) must be
able to replace it, causing the whole thing to get replaced up the tree.

Note that there may be other expression types that accept a table - not a column. They'd also have to use TableReferenceExpression.